app/pyproject.toml
[project]
name = "neurogram"
version = "0.1.0"
description = "Neurogram monorepo (api + workers)"
authors = [{ name = "Michael Chernyakoff", email = "chernyakoff@gmail.com" }]
dependencies = [
    "aerich>=0.9.1",
    "aioboto3>=15.1.0",
    "aioclock>=0.3.0",
    "aiohttp-socks>=0.10.1",
    "aiopath>=0.7.7",
    "anthropic>=0.74.0",
    "astor>=0.8.1",
    "autoflake>=2.3.1",
    "cryptg>=0.5.1",
    "cyclopts>=3.19.0",
    "emojis>=0.7.0",
    "fastapi>=0.116.1",
    "hatchet-sdk>=1.18.1",
    "html-to-markdown>=2.9.1",
    "httpx[socks]>=0.28.1",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.14.0",
    "openrouter>=0.1.1",
    "passlib[bcrypt]>=1.7.4",
    "phonenumbers>=9.0.13",
    "pydantic-settings-yaml>=0.2.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pysocks>=1.7.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "telethon>=1.41.2",
    "tiktoken>=0.12.0",
    "tortoise-orm[accel,asyncpg]>=0.25.1",
    "tortoise-serializer>=1.6.0",
    "uvicorn[standard]>=0.35.0",
    "watchfiles>=1.1.0",
]
requires-python = "==3.13.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false


[tool.taskipy.tasks]
cli = "python -m console"
migration = "aerich migrate"
migrate = "aerich upgrade"
unmigrate = "aerich downgrade"
autoflake = "autoflake ."


[tool.aerich]
tortoise_orm = "config.tortoise_config"
location = "./migrations"
src_folder = "."


[tool.autoflake]
ignore-init-module-imports = true
remove-all-unused-imports = true
recursive = true
in-place = true


[tool.workers.base]
name = "base"
slots = 20

[tool.workers.dialog]
name = "dialog"
slots = 20

[tool.sync]
workers_dir = "workers"
output_root = "api/hatchet"
client_import = "from api.hatchet.client import hatchet"
exclude = ["dialog"]


[dependency-groups]
dev = ["taskipy>=1.14.1"]
